cloud_storage: move partition_manifest to a separate file
