+++ src/v/cloud_storage/CMakeLists.txt
@@ -5,6 +5,7 @@ v_cc_library(
     cache_service.cc
     manifest.cc
     topic_manifest.cc
+    partition_manifest.cc
     recursive_directory_walker.cc
     remote.cc
     offset_translation_layer.cc
